feat: 사이드바 로그아웃 모달
