+++ becok/src/layout/HomeLayout.jsx
@@ -1,6 +1,5 @@
-// src/layout/HomeLayout.jsx
 import React from "react";
-import Sidebar from "../components/Sidebar";
+import Sidebar from "../components/Sidebar/Sidebar";
 import { Outlet } from "react-router-dom";
 import "./css/HomeLayout.css";
